Adapt to Teamscale 6.4.2
(Addressed TS-25288)
