+++ build.gradle
@@ -8,7 +8,7 @@ plugins {
 }
 
 ext {
-	teamscaleVersion = '5.9.0'
+	teamscaleVersion = '6.4.2'
 }
 
 test {
