+++ pom.xml
@@ -10,7 +10,7 @@
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<maven.compiler.source>1.8</maven.compiler.source>
 		<maven.compiler.target>1.8</maven.compiler.target>
-		<teamscaleVersion>5.9.0</teamscaleVersion>
+		<teamscaleVersion>6.4.2</teamscaleVersion>
 	</properties>
 
 	<dependencyManagement>
