+++ src/test/java/org/example/check/CheckTest.java
@@ -1,0 +1,29 @@
+/*-------------------------------------------------------------------------+
+|                                                                          |
+| Copyright 2005-2011 the ConQAT Project                                   |
+|                                                                          |
+| Licensed under the Apache License, Version 2.0 (the "License");          |
+| you may not use this file except in compliance with the License.         |
+| You may obtain a copy of the License at                                  |
+|                                                                          |
+|    http://www.apache.org/licenses/LICENSE-2.0                            |
+|                                                                          |
+| Unless required by applicable law or agreed to in writing, software      |
+| distributed under the License is distributed on an "AS IS" BASIS,        |
+| WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. |
+| See the License for the specific language governing permissions and      |
+| limitations under the License.                                           |
++-------------------------------------------------------------------------*/
+package org.example.check;
+
+import eu.cqse.check.framework.checktest.CheckTestBase;
+
+/**
+ * Main class for testing the checks in this library.
+ */
+public class CheckTest extends CheckTestBase {
+
+	public CheckTest() {
+		super(SampleCheck.class);
+	}
+}
