--- src/test/resources/eu/cqse/check/sample/SampleCheck/SampleCheckCode.java.expected
@@ -1,2 +1,0 @@
-Lines 3-3: This is a sample finding for method foo.
-Lines 6-6: This is a sample finding for method bar.
