--- src/test/resources/eu/cqse/check/sample/SampleCheckUsingPhase/Test1/Test1File1.java
@@ -1,5 +1,0 @@
-public class SampleCheckCode {
-	public class Inner {
-
-	}
-}
