--- src/test/resources/eu/cqse/check/sample/SampleCheckUsingPhase/Test1/Test1File1.java.expected
@@ -1,1 +1,0 @@
-Offsets 32-55 (lines 2-4): found another type with same name
