--- src/test/resources/eu/cqse/check/sample/SampleCheckUsingPhase/Test2/Test2File1.java
@@ -1,5 +1,0 @@
-public class SampleCheckCode {
-	public class Inner {
-		
-	}
-}
