--- src/test/resources/eu/cqse/check/sample/SampleCheckWithParameter/CustomExclude/SampleCheckCode.java
@@ -1,9 +1,0 @@
-public class SampleCheckCode {
-
-	private void foo(int a) {
-	}
-
-	private void bar(int a, int b) {
-	}
-
-}
