--- src/test/resources/eu/cqse/check/sample/SampleCheckWithParameter/ExcludedFile/SampleCheckCodeTest.java
@@ -1,9 +1,0 @@
-public class SampleCheckCodeTest {
-
-	private void foo(int a) {
-	}
-
-	private void bar(int a, int b) {
-	}
-
-}
