+++ src/test/resources/org/example/check/SampleCheck/SampleCheckCode.java
@@ -1,0 +1,11 @@
+package org.example.check.SampleCheck;
+
+public class SampleCheckCode {
+
+	private void foo(int a) {
+	}
+
+	private void bar(int a, int b) {
+	}
+
+}
