+++ src/test/resources/org/example/check/SampleCheck/SampleCheckCode.java.expected
@@ -1,0 +1,2 @@
+Lines 5-5: This is a sample finding for method foo.
+Lines 8-8: This is a sample finding for method bar.
