+++ src/test/resources/org/example/check/SampleCheckUsingPhase/Test1/Test1File1.java
@@ -1,0 +1,7 @@
+package org.example.check.SampleCheckUsingPhase.Test1;
+
+public class SampleCheckCode {
+	public class Inner {
+
+	}
+}
