+++ src/test/resources/org/example/check/SampleCheckUsingPhase/Test1/Test1File1.java.expected
@@ -1,0 +1,1 @@
+Offsets 88-111 (lines 4-6): found another type with same name
