+++ src/test/resources/org/example/check/SampleCheckUsingPhase/Test2/Test2File2.java
@@ -1,0 +1,7 @@
+package org.example.check.SampleCheckUsingPhase.Test2;
+
+public class SampleCheckCode2 {
+	public class Inner2 {
+		
+	}
+}
