+++ src/test/resources/org/example/check/SampleCheckWithParameter/CustomExclude/SampleCheckCode.java
@@ -1,0 +1,10 @@
+package org.example.check.SampleCheckWithParameter.CustomExclude;
+
+public class SampleCheckCode {
+
+	private void foo(int a) {
+	}
+
+	private void bar(int a, int b) {
+	}
+}
