+++ src/test/resources/org/example/check/SampleCheckWithParameter/ExcludedFile/SampleCheckCodeTest.java
@@ -1,0 +1,10 @@
+package org.example.check.SampleCheckWithParameter.ExcludedFile;
+
+public class SampleCheckCodeTest {
+
+	private void foo(int a) {
+	}
+
+	private void bar(int a, int b) {
+	}
+}
